Remove "Sixel" from "FINAL Scaled Score" entirely,
From feedback found on HN comments, it's not really fair to score by
sixel support, this is a legacy image protocol. It's pretty awful, but,
it's also the most commonly supported and easily detected. I might add
columns for their image protocols later so long as their support can be
automatically detected.